Clone element when necessary to inject key
We're re-creating any object that doesn't have a key and injecting a key
into it, which is causing non-react elements such as arrays and
`expect.any` style matchers to be mutated into a shape that can't then
be used.

Arrays are valid as children and should be preserved, and expect.any is
useful for matching translations.

We should be using React's cloneElement when we need to mutate an
element in this way so that it can still be tested against when
injecting a key prop into it.